+++ src/backends/sdl_window.cpp
@@ -230,7 +230,7 @@ void SdlWindow::render_snapshot(const SceneSnapshot& snap, TTF_Font* font, const
         SDL_SetRenderViewport(renderer_, &viewport);
     }
 
-    for (uint16_t idx : snap.z_order) {
+    for (uint32_t idx : snap.z_order) {
         render_draw_list(*snap.draw_lists[idx], font);
     }
     if (!snap.overlay.empty()) {
